requirements.txt
pyyaml==6.0.1
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.4.2
pandas==2.1.2
pymongo==4.10.1
#dagger_python
pip~=24.3.1
dagger-io~=0.15.1
anyio~=4.3.0
requests~=2.32.3
kcl_lib==0.10.8